[Movement] Fix the new orientation computation
Updated the angle function. It now returns an integer between 0 and 359
(no negative number). The new orientation computation is made so that
it is symmetric. Turning right or left had rounding problems. The new
logic ensures that the rounding problems are solved by taking the rotation
wich causes the ship to rotate less.